backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
python-dotenv==1.0.1
anthropic>=0.40.0
gunicorn
stripe>=8.0.0
resend>=0.7.0
Pillow>=10.0.0
